lower timeout so shift-O works faster
https://github.com/vim/vim/issues/24

On many terminals, <Esc>O (with an uppercase O) is a prefix for several
keycodes: to see them, do

    :set termcap

in Vim running on the terminal where you experience the problem, and notice
that several code sequences begin with ^[O where the blue ^[ means Esc.

I recommend the following setting:

    :set timeout timeoutlen=5000 ttimeoutlen=100

where the timeouts are in milliseconds and should be chosen as follows:

timeoutlen: longer than the time it takes you to type successive characters of
a multi-character mapping or abbreviation at your slowest typing speed, but not
too much longer than that, to avoid making you impatient when waiting for a
timeout

ttimeoutlen: shorter than the time it takes you to hit successive keys at your
fastest typing speed, but longer that the maximum time between successive bytes
sent by the keyboard driver for a single (special) key.

See
    :help 'timeout'
    :help 'timeoutlen'